mejorado y corregido el digest response

+++ uaclient.py
@@ -1,7 +1,7 @@
 #!/usr/bin/python3
 # -*- coding: utf-8 -*-
 # Made by Felipe Sandoval Sibada
-"""UA Client Program that sends SIP methods request."""
+"""UA SIP Client Program that sends RTP Audio."""
 
 from xml.sax import make_parser
 from xml.sax.handler import ContentHandler
@@ -10,6 +10,7 @@ import socket
 import sys
 import random
 import os
+import hashlib
 
 
 try:
@@ -29,14 +30,14 @@ except (IndexError, ValueError):
 
 
 class ConfigHandler(ContentHandler):
-    """For handling Configuration entries"""
+    """For handling configuration entries in XML format type."""
 
     def __init__(self):
-        """Making a list with my configuration"""
+        """Making a list with my configuration."""
         self.myconfig = []
 
     def startElement(self, name, attr):
-        """Método que se llama cuando se abre una etiqueta"""
+        """Method to get data from my ATTLISTS."""
         if name == "account":       # one way to do it
             username = attr.get('username', "")
             self.myconfig.append(username)
@@ -63,8 +64,9 @@ class ConfigHandler(ContentHandler):
         elif name == 'audio':
             path_audio = attr.get('path', "")
             self.myconfig.append(path_audio)
-    
+
     def get_config(self):
+        """My configurations setting list."""
         return self.myconfig
 
 
@@ -73,6 +75,19 @@ def actual_time():
     timenow = time.strftime("%Y%m%d%H%M%S", time.gmtime(time.time()))
     return timenow
 
+
+def sents_log(config, log_file, sip_data):
+    """Sent content for log chronology purposes."""
+    log_file.write(str(actual_time()) + " Sent to " + config[5] +
+                   ":" + config[6] + ": " + sip_data + "\n")
+
+
+def recieved_log(config, log_file, sip_data):
+    """Recieved content for log chronology purposes."""
+    log_file.write(str(actual_time()) + " Recieved from " + config[5] +
+                   ":" + config[6] + ": " + sip_data + "\n")
+
+
 if __name__ == "__main__":
     try:
         parser = make_parser()
@@ -80,69 +95,78 @@ if __name__ == "__main__":
         parser.setContentHandler(cHandler)
         parser.parse(open(CONFIG))
         config = cHandler.get_config()
+        if not os.path.exists(config[-1]):  # Does this audio file exists?.
+            raise OSError
+
         try:
             log_file = open(config[7])
             log_file = open(config[7], "a")
-        except FileNotFoundError:   #  When the file does not exists.
+        except FileNotFoundError:   # When the file does not exists.
             log_file = open(config[7], "w")
             log_file.write(str(actual_time()) + " Starting...\n")
-# Different kind of methods.
+
         if METHOD == "REGISTER" or METHOD == "register":
+            SIP_LINE = METHOD + " sip:" + config[0] + ":" + config[3] +\
+                       " SIP/2.0\r\n\r\n" + "Expires: "
             if len(sys.argv) == 4:
-                SIP_LINE = METHOD + " sip:" + config[0] + ":" +\
-                           config[3] + " SIP/2.0\r\n\r\n" + "Expires: " +\
-                           OPTION + "\r\n"
+                SIP_LINE += OPTION + "\r\n"
             else:
-                SIP_LINE = METHOD + " sip:" + config[0] + ":" +\
-                           config[3] + " SIP/2.0\r\n\r\n" + "Expires: 3600\r\n"
+                SIP_LINE += "3600\r\n"      # Default expiration time.
         else:
             if len(sys.argv) == 4:
-                if METHOD == "INVITE":
-                    SIP_LINE = METHOD + " sip:" + OPTION + " SIP/2.0\r\n\r\n"\
-                               + "Content-Type: application/sdp\r\n\r\n" +\
-                               "v=0\r\n" + "o=" + config[0] + " " + config[2]\
-                               + "\r\n" + "s=PracticaFinal\r\n" + "t=0\r\n" +\
-                               "m=audio " + config[4] + " RTP\r\n"
-                else:
-                    SIP_LINE = METHOD + " sip:" + OPTION + " SIP/2.0\r\n\r\n"
+                SIP_LINE = METHOD + " sip:" + OPTION + " SIP/2.0\r\n"
+                if METHOD == "INVITE" or METHOD == "invite":
+                    SIP_LINE += "Content-Type: application/sdp\r\n\r\n" +\
+                                "v=0\r\n" + "o=" + config[0] + " " + config[2]\
+                                + "\r\n" + "s=PracticaFinal\r\n" + "t=0\r\n" +\
+                                "m=audio " + config[4] + " RTP\r\n"
             else:
-                print("FALTA SABER EL LOGIN")
+                print("-- IMPORTANT: you did't specify the login --\n")
                 SIP_LINE = METHOD + " SIP/2.0\r\n\r\n"
 
-        SIP_HASH = SIP_LINE.split("\r\n")
+        SIP_HASH = (" ").join(SIP_LINE.split())     # Content to write in log.
+
         with socket.socket(socket.AF_INET, socket.SOCK_DGRAM) as my_socket:
             my_socket.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
             my_socket.connect((config[5], int(config[6])))
             print(SIP_LINE)     # Content to send.
-
             my_socket.send(bytes(SIP_LINE, 'utf-8'))
-            log_file.write(str(actual_time()) + " Sent to " + config[5] +
-                           ":" + config[6] + ": " + SIP_HASH[0] + " " + 
-                           SIP_HASH[2] + "\n")
+            sents_log(config, log_file, SIP_HASH)
             data = my_socket.recv(1024)
+
             print('-- RECIEVED SIP RESPONSES --\n' + data.decode('utf-8'))
-            data_hash = data.decode('utf-8').split(" ")
-            log_file.write(str(actual_time()) + " Recieved from " + config[5] + 
-                           ":" + config[6] + ": " + data_hash[0] + "\n")
+            data_hash = data.decode('utf-8').split()
+            log_data = (" ").join(data_hash)        # Content to write in log.
+            recieved_log(config, log_file, log_data)
+            nonce_recieved = data_hash[6].split("=")[1]
             if "401" in data_hash:
+                digest = hashlib.md5()
+                digest.update(bytes(nonce_recieved, "utf-8"))
+                digest.update(bytes(config[1], "utf-8"))
+                digest.digest
                 response = random.randint(000000000000000000000,
                                           999999999999999999999)
-                SIP_LINE += "Authorization: Digest response=" + str(response)\
-                            + "\r\n"
+                SIP_LINE += "Authorization: Digest response=" +\
+                            digest.hexdigest() + "\r\n"
+                SIP_HASH = (" ").join(SIP_LINE.split())
+                sents_log(config, log_file, SIP_HASH)
                 my_socket.send(bytes(SIP_LINE, 'utf-8'))
-                print("-- SENDING REGISTER AGAIN --\n" + SIP_LINE) 
-            elif data.decode('utf-8').split(" ")[-1] == "OK\r\n\r\n" and \
-               METHOD != "BYE":
+                print("-- SENDING REGISTER AGAIN --\n" + SIP_LINE)
+
+            elif "OK" in data_hash and METHOD != "BYE":
                 SIP_ACK = "ACK" + " sip:" + config[0] + " SIP/2.0\r\n\r\n"
+                # CORREGIR EL SIP DE ARRIBA. ES AL QUE SE LO ENVIO.
                 my_socket.send(bytes(SIP_ACK, 'utf-8'))
-                data = my_socket.recv(1024)
-                #send = "mp32rtp -i 127.0.0.1 -p 23032 < " + sys.argv[3]
-                #os.system(send)
-                print(data.decode('utf-8'))
+                # send = "mp32rtp -i 127.0.0.1 -p 23032 < " + sys.argv[3]
+                # os.system(send)
+            elif METHOD == "BYE":
+                log_file.write(str(actual_time()) + " Finishing.\n")
             my_socket.close()
             print("END OF SOCKET")
-            log_file.write(str(actual_time()) + " Finishing.\n")
+            # log_file.write(str(actual_time()) + " Finishing.\n")
             log_file.close()
+    except (FileNotFoundError, OSError):
+        sys.exit("Usage: python uaclient.py config method option.")
     except ConnectionRefusedError:
         log_file.write(str(actual_time()) + " Error: No server listening at " +
                        config[5] + " port " + config[6] + "\n")
